Fix pointer arithmetic errors in MonotonicAllocationStrategy

+++ src/strategy/Pool.cpp
@@ -19,10 +19,10 @@ Pool::Pool(
 
   UMPIRE_LOG("Creating " << m_slots << "-slot pool.");
 
-  m_lengths = new int[m_slots];
+  m_lengths = new size_t[m_slots];
   m_pointers = new void*[m_slots];
 
-  for (int i = 0; i < m_slots; ++i) {
+  for (size_t i = 0; i < m_slots; ++i) {
     m_pointers[i] = nullptr;
     m_lengths[i] = 0;
   }
@@ -33,7 +33,7 @@ Pool::allocate(size_t bytes)
 {
   void* ptr = nullptr;
 
-  for (int i = 0; i < m_slots; ++i) {
+  for (size_t i = 0; i < m_slots; ++i) {
      if (m_lengths[i] == bytes) {
         m_lengths[i] = -m_lengths[i] ;
         ptr = m_pointers[i] ;
@@ -52,7 +52,7 @@ Pool::allocate(size_t bytes)
 void
 Pool::deallocate(void* ptr)
 {
-  for (int i = 0; i < m_slots; ++i) {
+  for (size_t i = 0; i < m_slots; ++i) {
     if (m_pointers[i] == ptr) {
       m_lengths[i] = -m_lengths[i];
       ptr = nullptr;
